Include appropriate headers
In this project, class declarations frequently use the exchange() function,
which was added in C++17 and allows programmers a better way to make use of the
move semantic. However, the STL header file that corresponds to it is "utility"
rather than "algorithm", therefore this commit fixes the header and source
files that use exchange().